\name{CPVS-package}
\alias{CPVS-package}
\alias{CPVS}
\docType{package}
\title{
\packageTitle{CPVS}
}
\description{
\packageDescription{CPVS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CPVS}
\packageIndices{CPVS}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{CPVS}

Maintainer: \packageMaintainer{CPVS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
