% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_segments.R
\name{fit_segments}
\alias{fit_segments}
\title{Estimate the signal's parameters in each segment given a set of change points}
\usage{
fit_segments(y, Xsin, Xcos, Ne, cp, sigma2 = 0, pl = TRUE)
}
\arguments{
\item{y}{vector of observations}

\item{Xsin}{matrix of sine functions X^(1) based on the grid}

\item{Xcos}{matrix of cosine functions X^(2) based on the grid}

\item{Ne}{number of effects in the SuSiE algorithm (can be different from the one used for segmentation, if a better estimate of the signal is desired)}

\item{cp}{estimated change points from the proposed procedure}

\item{sigma2}{default is 0, meaning that the variance of the error term is estimated. It can be fixed to a specific value.}

\item{pl}{if \code{TRUE}, plot the estimates over the observed values}
}
\value{
\code{models}: list of models estimated in each segment (output of \code{SuSiE_group})

\code{signal_estimate}: estimate of the signal, given the change points
}
\description{
Estimate the signal's parameters in each segment given a set of change points
}
