% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Xmatrices.R
\name{getX}
\alias{getX}
\title{Get the design matrices X^(1) and X^(2)}
\usage{
getX(x, om)
}
\arguments{
\item{x}{vector of times}

\item{om}{the grid of frequencies}
}
\value{
\code{Xsin}: matrix X^(1)

\code{Xcos}: matrix X^(2)
}
\description{
Get the design matrices X^(1) and X^(2)
}
