% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_cp.R
\name{pruneSet}
\alias{pruneSet}
\title{Prune the set of candidate change points using an information criterion (IC) (see paper for details)}
\usage{
pruneSet(
  cp_set,
  y,
  Xsin,
  Xcos,
  Ne,
  sigma2 = 0,
  IC = "MDL",
  option = "best IC",
  Trace = 0,
  pl = FALSE,
  re_test = 3,
  fixCP = 0,
  maxEval = 8
)
}
\arguments{
\item{cp_set}{output of \code{getCandidateSet}}

\item{y}{vector of observations}

\item{Xsin}{matrix of sine functions X^(1) based on the grid}

\item{Xcos}{matrix of cosine functions X^(2) based on the grid}

\item{Ne}{number of effects in the SuSiE algorithm}

\item{sigma2}{default is 0, meaning that the variance of the error term is estimated. It can be fixed to a specific value.}

\item{IC}{the information criterion to be used. "MDL" (minimum description length) and "mBIC" modified BIC are possible}

\item{option}{the type of pruning to perform (more details...)}

\item{Trace}{if 1, print progress of the pruning}

\item{pl}{if \code{TRUE}, plot the IC path}
}
\value{
The estimated set of change points
}
\description{
Prune the set of candidate change points using an information criterion (IC) (see paper for details)
}
